% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.possible_n}
\alias{autoplot.possible_n}
\title{Autoplot estimate count sum}
\usage{
\method{autoplot}{possible_n}(x, row = 1, threshold = 1, n_lim, by_taxa = TRUE, show_score = FALSE)
}
\arguments{
\item{x}{output of estimate count sum}

\item{row}{integer giving name of row with sample to plot}

\item{threshold}{numeric giving minimum proportion of taxa that
need to be concordant with count sum.}

\item{n_lim}{Limits on range of n.
Defaults to full range available which can cause problems when by_taxa is TRUE
as lines may fail to show because of screen resolution.}

\item{by_taxa}{logical giving name of column with taxon names}

\item{show_score}{logical. Show scores instead of main plot}
}
\value{
ggplot object.
}
\description{
Plots the direct search attempt to estimate the count sum from percent data
}
\details{
With \code{show_score == TRUE}, a plot of the proportion of taxa with percent
abundances concordant with each count sum is shown.
With \code{by_taxa == TRUE}, shows a plot of with taxa have percent abundances
concordant with each count sum.
With a wide range of possible count sums, the lines drawn become thinner
than screen resolution, so use the \code{n_lim} option to avoid this.
With \code{by_taxa == FALSE}, taxa with the same abundance are grouped together.
A horizontal line is drawn at each putative count sum where the proportion
of concordant taxa is \verb{> threshold}.
Small \code{x} are plotted for non-concordant taxa when \code{threshold < 1}.
This is probably the most useful plot.
}
\examples{
require("dplyr")
require("ggplot2")
data(last_chance)
last_chance_est <- last_chance0  \%>\%
  tidyr::pivot_longer(cols = -c(age_calBP, totcaps),
                      names_to = "taxon", values_to = "percent") \%>\%
  estimate_n(digits = 2, ID_cols = c("age_calBP", "totcaps"))
autoplot(last_chance_est, row = 10, by_taxa = FALSE)
autoplot(last_chance_est, row = 10, show_score = TRUE)

}
\references{
Telford (2019) Tools for identifying unexpectedly low
microfossil count sums. Preprint.
}
