% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est2_n.R
\name{estimate_n}
\alias{estimate_n}
\title{Estimate count sum}
\usage{
estimate_n(x, percent_col = "percent", taxon_col = "taxon", ID_cols,
  digits = 2, nmin = 1L, nmax = 1000L)
}
\arguments{
\item{x}{data.frame of percent data in long format}

\item{percent_col}{character giving name of column with percent data}

\item{taxon_col}{character giving name of column with taxon names}

\item{ID_cols}{character vector giving name of one or more columns with sample IDs}

\item{digits}{numeric; how many decimal places are the percent reported to}

\item{nmin}{integer; possible count sum to start direct search at}

\item{nmax}{integer; possible count sum to end direct seach at}
}
\description{
Estimate count sum from percent data
}
\examples{
require("dplyr")
data(last_chance)
last_chance0  \%>\% 
  tidyr::gather(key = taxon, value = percent, -age_calBP, -totcaps) \%>\% 
  estimate_n(digits = 2, ID_cols = c("age_calBP", "age_calBP"))

}
