% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_chance.R
\docType{data}
\name{last_chance}
\alias{last_chance}
\alias{last_chance0}
\title{Chironomid assemblages from Last Chance Lake.}
\format{
A data frame with 22 rows and 28 variables:
\describe{
  \item{age_calBP}{Sample age}
  \item{totcaps}{Count sum}
  \item{Corynon}{First species}
  ...
}
}
\source{
\url{https://www.ncdc.noaa.gov/paleo-search/study/22091}
}
\usage{
data(last_chance)
}
\description{
A dataset containing the ages, counts sums and chironomid assemblages from 22 samples from Last Chance Lake from Axford et al (2017).
}
\examples{
data(last_chance)
}
\references{
Axford, Y. et al. 2017. Timing and magnitude of early to middle Holocene warming in East Greenland inferred from chironomids. Boreas. doi: 10.1111/bor.12247
}
\keyword{datasets}
