% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_check_functions.R
\name{percent_checker}
\alias{percent_checker}
\title{Percent checker}
\usage{
percent_checker(spp, digits, site_column, count_column)
}
\arguments{
\item{spp}{data.frame of species percent data, optionally with columns for site and count sum}

\item{digits}{integer giving the precision of the percent data}

\item{site_column}{character giving name of column with site/sample names. If missing uses rownames instead.}

\item{count_column}{character giving name of column with count sums. If missing count sum is estimated from smallest percentage.}
}
\description{
check percent in species file are plausible
}
\examples{
data(last_chance)
last_chance <- dplyr::select(last_chance0, -totcaps)
percent_checker(spp = last_chance, digits = 2, site_column = "age_calBP")
}
