% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_check_functions.R
\name{estimate_n}
\alias{estimate_n}
\title{Count estimator}
\usage{
estimate_n(spp, digits)
}
\arguments{
\item{spp}{data.frame of species percent data}

\item{digits}{integer giving precision of species data}
}
\description{
Estimates count sums of species percent data
}
\examples{
require("dplyr")
data(last_chance)
last_chance <- select(last_chance0, -age_calBP, -totcaps)
estimate_n(spp = last_chance, digits = 2)
}
